+++ Examples/WinForms/Cartesian/Labels/Labels.resx
